pyproject.toml
[project]
name = "semantic-judge"
version = "0.1.0"
description = "Compares sentences for similarity"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "sentence-transformers>=5.1.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
]
